Don't repack kernels into images that never had a kernel
If the image to be repacked doesn't have a kernel in it, then it should not
have a kernel shoehorned into the new image. This issue would cause
magiskboot to generate broken vendor images upon repacking when a kernel
file is present in the same directory as the components to be repacked.

Fix it by never repacking a kernel into an image that didn't have one.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>